Fixed a bug with autostarting files that load up to FFFF.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@22833 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/autostart-prg.c
@@ -49,8 +49,8 @@ static autostart_prg_t *inject_prg;
 
 static autostart_prg_t * load_prg(const char *file_name, fileio_info_t *finfo, log_t log)
 {
-    WORD ptr;
-    int  end;
+    DWORD ptr;
+    DWORD end;
     BYTE lo,hi;
     int  i;
     autostart_prg_t *prg;
